+++ hoot-services/src/main/java/hoot/services/controllers/osm/MapResource.java
@@ -574,13 +574,13 @@ public class MapResource
       final JSONObject anode = currMap.retrieveANode(queryBounds);
       long nodeCnt = currMap.getNodesCount(queryBounds);
 
-      double dMinLon = DbUtils.fromDbCoordValue((Integer) extents.get("minlon"));
-      double dMaxLon = DbUtils.fromDbCoordValue((Integer) extents.get("maxlon"));
-      double dMinLat = DbUtils.fromDbCoordValue((Integer) extents.get("minlat"));
-      double dMaxLat = DbUtils.fromDbCoordValue((Integer) extents.get("maxlat"));
+      double dMinLon = DbUtils.fromDbCoordValue((Long) extents.get("minlon"));
+      double dMaxLon = DbUtils.fromDbCoordValue((Long) extents.get("maxlon"));
+      double dMinLat = DbUtils.fromDbCoordValue((Long) extents.get("minlat"));
+      double dMaxLat = DbUtils.fromDbCoordValue((Long) extents.get("maxlat"));
 
-      double dFirstLon = DbUtils.fromDbCoordValue((Integer) anode.get("lon"));
-      double dFirstLat = DbUtils.fromDbCoordValue((Integer) anode.get("lat"));
+      double dFirstLon = DbUtils.fromDbCoordValue((Long) anode.get("lon"));
+      double dFirstLat = DbUtils.fromDbCoordValue((Long) anode.get("lat"));
 
       ret.put("minlon", dMinLon);
       ret.put("maxlon", dMaxLon);
